Only clear temporary directory if it already exists.

+++ test/test_capsules.py
@@ -1,3 +1,4 @@
+import os
 import shutil
 from unittest import TestCase
 
@@ -5,7 +6,8 @@ from capsules.capsules import capsules
 
 # Remove directory if it exists to ensure
 # tests have full coverage.
-shutil.rmtree('./__capsules__')
+if os.path.exists('./__capsules__'):
+    shutil.rmtree('./__capsules__')
 
 # Definitions of functions used in unit tests.
 identity = lambda x: x
